pyproject.toml
[project]
name = "python-1-pillow-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.0.0",
    "python-dotenv>=1.0.1",
]
